[GODMUSIC-7427] [Server] FLO WEB 관련 수정사항 적용 : 홈패널 아이디를 웹인 경우 String 으로 리턴

+++ src/main/java/com/sktechx/godmusic/personal/rest/service/impl/recommend/phase/PersonalRecommendPhaseServiceImpl.java
@@ -187,7 +187,7 @@ public class PersonalRecommendPhaseServiceImpl  implements PersonalRecommendPhas
         }).map(panel->{
             PanelContentVo content = panel.getContent();
             if(content != null && content.getId() != null){
-                return content.getId();
+                return (Long) content.getId();
             }
             return null;
         }).collect(Collectors.toList());
